requirements.txt
numpy~=1.20.1
scipy~=1.6.1
qiskit~=0.27.0
pytorch~=1.9.1